Move SnsrNotice to datatypes.py
