+++ src/qtpy_datalogger/datatypes.py
@@ -4,7 +4,13 @@ Shared constants and classes used by the host.
 This module uses CPython features not available in the CircuitPython implementation.
 """
 
+import datetime
 import enum
+import pathlib
+import subprocess
+from typing import Any, NamedTuple
+
+import toml
 
 
 class Links(enum.StrEnum):
@@ -44,3 +50,47 @@ class SnsrPath(enum.StrEnum):
 
     root = "snsr"
     notice = "snsr/notice.toml"
+
+
+class SnsrNotice(NamedTuple):
+    """Represents the contents of the notice.toml file for a sensor_node."""
+
+    comment: str
+    version: str
+    commit: str
+    timestamp: datetime.datetime
+
+    @staticmethod
+    def get_package_notice_info(allow_dev_version: bool) -> "SnsrNotice":
+        """Detect and generate the information used in the notice.toml file."""
+        this_file = pathlib.Path(__file__)
+        this_folder = this_file.parent
+        notice_toml = this_folder.joinpath("sensor_node", SnsrPath.notice)
+        notice_contents = toml.load(notice_toml)
+        snsr_notice = SnsrNotice(**notice_contents)
+        my_comment = snsr_notice.comment
+        my_version = snsr_notice.version
+        my_commit = snsr_notice.commit
+        my_timestamp = snsr_notice.timestamp
+
+        if __package__:
+            # We're installed
+            import importlib.metadata
+
+            my_version = importlib.metadata.version(str(__package__))
+
+        # When we're running from the git source, we're in development mode
+        this_package_parent = this_file.parent.parent
+        in_dev_mode = this_package_parent.name == "src"
+        if in_dev_mode:
+            if allow_dev_version:
+                my_version = f"{my_version}.post0.dev0"
+
+            most_recent_commit_info = ["git", "log", "--max-count=1", "--format=%h %aI"]
+            sha_with_timestamp = subprocess.check_output(most_recent_commit_info).strip()  # noqa: S603 -- command is well-formed and user cannot execute arbitrary code
+            sha_and_timestamp = sha_with_timestamp.decode("UTF-8").split(" ")
+            my_commit = sha_and_timestamp[0]
+            my_timestamp = datetime.datetime.fromisoformat(sha_and_timestamp[1])
+
+        my_comment = f"Generated by '{__name__}.py'"
+        return SnsrNotice(my_comment, my_version, my_commit, my_timestamp)
