+++ src/qtpy_datalogger/equip.py
@@ -19,8 +19,9 @@ import findimports
 import packaging.version
 import toml
 
-from . import discovery
-from .datatypes import ExitCode, Links, SnsrPath
+from qtpy_datalogger import discovery
+
+from .datatypes import ExitCode, Links, SnsrNotice, SnsrPath
 
 logger = logging.getLogger(__name__)
 
@@ -36,15 +37,6 @@ class Behavior(StrEnum):
     Compare = "Compare"
     Describe = "Describe"
     Force = "Force"
-
-
-class SnsrNotice(NamedTuple):
-    """Represents the contents of the notice.toml file for a sensor_node."""
-
-    comment: str
-    version: str
-    commit: str
-    timestamp: datetime.datetime
 
 
 class SnsrNodeBundle(NamedTuple):
@@ -136,7 +128,7 @@ def _detect_snsr_bundle(main_folder: pathlib.Path) -> SnsrNodeBundle:
 
     detect_message = f"Probing for sensor_node at '{main_folder}'"
     if detecting_self:
-        snsr_notice = _get_package_notice_info(allow_dev_version=True)
+        snsr_notice = SnsrNotice.get_package_notice_info(allow_dev_version=True)
         detect_message = ""
     elif notice_file.exists():
         file_contents = notice_file.read_text()
@@ -340,49 +332,9 @@ def _equip_snsr_node(behavior: Behavior, comparison_information: dict[str, SnsrN
 
 def _create_notice_file_contents(allow_dev_version: bool) -> str:
     """Format the notice.toml file for the package."""
-    snsr_notice = _get_package_notice_info(allow_dev_version)
+    snsr_notice = SnsrNotice.get_package_notice_info(allow_dev_version)
     file_contents = toml.dumps(snsr_notice._asdict())
     return file_contents
-
-
-def _get_package_notice_info(allow_dev_version: bool) -> SnsrNotice:
-    """Detect and generate the information used in the notice.toml file."""
-    logger.debug("Getting notice information from notice file")
-
-    this_file = pathlib.Path(__file__)
-    this_folder = this_file.parent
-    notice_toml = this_folder.joinpath("sensor_node", SnsrPath.notice)
-    notice_contents = toml.load(notice_toml)
-    snsr_notice = SnsrNotice(**notice_contents)
-    my_comment = snsr_notice.comment
-    my_version = snsr_notice.version
-    my_commit = snsr_notice.commit
-    my_timestamp = snsr_notice.timestamp
-
-    if __package__:
-        # We're installed
-        import importlib.metadata
-
-        logger.debug("Updating version from __package__ metadata")
-        my_version = importlib.metadata.version(str(__package__))
-
-    # When we're running from the git source, we're in development mode
-    this_package_parent = this_file.parent.parent
-    in_dev_mode = this_package_parent.name == "src"
-    if in_dev_mode:
-        logger.debug("Updating notice information from development environment")
-        if allow_dev_version:
-            logger.debug("Including the version")
-            my_version = f"{my_version}.post0.dev0"
-
-        most_recent_commit_info = ["git", "log", "--max-count=1", "--format=%h %aI"]
-        sha_with_timestamp = subprocess.check_output(most_recent_commit_info).strip()  # noqa: S603 -- command is well-formed and user cannot execute arbitrary code
-        sha_and_timestamp = sha_with_timestamp.decode("UTF-8").split(" ")
-        my_commit = sha_and_timestamp[0]
-        my_timestamp = datetime.datetime.fromisoformat(sha_and_timestamp[1])
-
-    my_comment = f"Generated by '{__name__}.py'"
-    return SnsrNotice(my_comment, my_version, my_commit, my_timestamp)
 
 
 def _get_plugins(folder_list: list[pathlib.Path]) -> list[str]:
