+++ tests/test_console.py
@@ -64,7 +64,7 @@ def test_generate_notice_option(tmp_path: pathlib.Path) -> None:
     assert result.exit_code == 0
     assert output_file.exists()
     all_lines = output_file.read_text().splitlines()
-    assert all_lines[0] == "comment = \"Generated by 'qtpy_datalogger.equip.py'\""
+    assert all_lines[0] == "comment = \"Generated by 'qtpy_datalogger.datatypes.py'\""
     assert all_lines[-1].startswith("timestamp = ")
